requirements.txt
./deployment/anomaly_detection/ml-essentials
click
jinja2
networkx
numpy
pandas
PyYAML
python-snappy
scikit-learn
seaborn
semver
natsort
imageio
fs
lazy_object_proxy
kafka-python
tqdm
loguru
numba
tensorboard==1.15
flask
lru-dict
cffi
graphviz


-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.12.0

-f https://data.dgl.ai/wheels/repo.html
dgl
